The module chargetransfer has been redesigned for a new architecture. Event listeners and event publishers have been added. Now, the interaction logic based on event-driven architecture is encapsulated within the ChargeTransferBlock class, while the original logic of the class has been moved to ChargeTransferBlockTask. The ChargeTransferBlockInterface interface has been reworked.
Now, when calling the startCharging() method, the run method of the ChargeTransferBlockTask class will be executed by the executor in a separate thread. After the task completes, one of the implementations of StopChargingEventPublisher will generate an event indicating the end of charging (the event type depends on the shutdown initiator).