During the handling of the premature emergency disconnection case, the architecture was expanded. A new state, Unavailable, was added, which is used when the electric vehicle (EV) disconnects, but other EVs cannot yet connect to the station. This state can be entered using the GET_UNAVAILABLE action and can be transitioned into from any other state. Transitioning to the Available state is now only possible from the Unavailable state.
The API of the commev module was extended, and now the EVCommunicationBlockInterface provides new methods - availableConnection() and acceptDisconnect(). Now, each time before sending a SessionStopRes message, the acceptDisconnect method needs to be called, and to allow EVs to reconnect to the charging station, the availableConnection() method needs to be invoked. For control over the Unavailable and Available states, a new module, ConnectorManager, was created. It provides methods unlockConnector() and requestStopChargingEmergency().

The StateOperator, after the EV disconnection, defaults to changing the state to UNAVAILABLE, and ConnectorManager is responsible for unlocking. In all cases except one, when handling the DisconnectedEV event, this module's listener invokes the unlockConnector method, but if disconnect permission was not granted and the charging transfer process is not currently stopped, this class calls the requestStopChargingEmergency() method. In this method, the StopChargingEmergency event is published, and during its handling, ChargeTransferBlock must urgently complete the task execution and then publish the ChargingStoppedEmergency event. When listening to this event, the ConnectorManager class unlocks the connector (after the charging process is completely stopped). Additionally, a parent ChargingStopped event was added for the ChargingStoppedNormally and ChargingStoppedEmergency event classes, so that different modules could subscribe to the charging stoppage, regardless of how it ended.